quickpars: Initial working version
This commit is an initial working version of an event driven parser for QuickJS bytecode, heavily inspired by wasm-tool's wasmparser. 

The general idea of this parser is to slice the bytecode into sections, and avoid constructing any intermediate representations eagearly, instead any sections that involve a "collection" of items are returned as-is to the caller via the `Payload`.