First step towards the runtime implementation
This commit is the first step towards the runtime implementation a.k.a
the opcode handler. The first iteration of the compiler aims to
delegate most of the operations to the JS runtime to eliminate the
interpreter loop.

One of the very first things that I am aiming to do with this change
is figuring out a way to create callable Wasm functions from the JS
engine by means of re-using the existing `JSCFunctionData` already
available in QuickJS. This commit is still not complete, is pretty
much the setup, working towards having callable functions.